Create text-based logic event reporter
The animation systems to come will need to be able to repond to logic
system events, and this text system is the first foray into a system
that processes such events to report to the player.

I also made an update to the logic to remove target translation because
it served no UX benefit and made logic more confusing. Symbolically this
allows for the metaphor of a SPOT as a lane or circuit, the latter in
particular which may lead to some helful design options.